图片抖动的原因： 通过3D绘制时，按缩放，宽度正常 通过直接光栅化时，绘制宽度比实际少1个像素 (x=60,y=70)坐标 实际是 (w=59,h=69) +1即可

+++ game/GameWind.h
@@ -21,7 +21,7 @@ private:
     HINSTANCE mhInstance = nullptr;
     static const wchar_t* wndClassName;
 public:
-    GameWind(HINSTANCE hInstacne, int width, int height);
+    GameWind(HINSTANCE hInstacne, float width, float height);
     GameWind(const GameWind& wnd);
     ~GameWind();
     void ShowWind();
@@ -30,4 +30,4 @@ public:
     static LRESULT CALLBACK WindowProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
 };
 
-GameWind* CALL windCreate(HINSTANCE hInstance, int width, int hegiht);
+GameWind* CALL windCreate(HINSTANCE hInstance, float width, float hegiht);
